feat: NSPersistentContainer를 생성.
- SceneDelegate에서 생성 후 DiaryViewController에 주입, 이후 화면 전환이 일어날 때 해당 뷰 컨트롤러에 주입해서 같은 container를 사용할 수 있도록 함
- 전역으로 Container를 갖게 되면 CoreData의 변경사항이 일어났을 때 어디서 변경이 일어났는지 정확히 파악하기 힘듬
- container만 가지고 있으면 사용할 수 있도록 Extension으로 PersistentContainer가 사용하는 메서드들을 분리
