+++ Diary/SceneDelegate.swift
@@ -4,16 +4,25 @@
 //  Copyright © yagom. All rights reserved.
 // 
 
+import CoreData
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
+    let persistentContainer: NSPersistentContainer = {
+        let container = NSPersistentContainer(name: "Diary")
+        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error (error), (error.userInfo)")
+            }
+        })
+        return container
+    }()
+    
     var window: UIWindow?
-
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let scene = scene as? UIWindowScene else { return }
-        let rootViewController = DiaryViewController()
+        let rootViewController = DiaryViewController(with: self.persistentContainer)
         let navigationController = UINavigationController(rootViewController: rootViewController)
         
         window = UIWindow(frame: scene.coordinateSpace.bounds)
@@ -50,9 +59,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // to restore the scene back to its current state.
 
         // Save changes in the application's managed object context when the application transitions to the background.
-        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+        self.persistentContainer.saveContext()
     }
-
-
 }
-
